server_code/requirements.txt
fastapi==0.104.0
firebase_admin==6.2.0
gradio_client==0.6.1
matplotlib==3.7.2
openai==0.28.1
Pillow==10.1.0
python-dotenv==1.0.0
Requests==2.31.0
transformers==4.31.0
uvicorn==0.23.2
